app/requirements.txt
fastapi
uvicorn[standard]
google-adk[extensions]>=1.0.0
google-generativeai
mcp
httpx
pydantic-settings